feat: :art: split workspace
